+++ backend/src/main/java/ddangkong/facade/admin/dto/AdminLoginRequest.java
@@ -1,4 +1,4 @@
-package ddangkong.controller.admin.dto;
+package ddangkong.facade.admin.dto;
 
 public record AdminLoginRequest(String nickname, String password) {
 }
