+++ backend/src/test/java/ddangkong/controller/admin/BaseAdminControllerTest.java
@@ -1,7 +1,7 @@
 package ddangkong.controller.admin;
 
 import ddangkong.controller.BaseControllerTest;
-import ddangkong.controller.admin.dto.AdminLoginRequest;
+import ddangkong.facade.admin.dto.AdminLoginRequest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
