+++ backend/src/test/java/ddangkong/facade/admin/AdminServiceTest.java
@@ -1,10 +1,11 @@
-package ddangkong.service.admin;
+package ddangkong.facade.admin;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import ddangkong.exception.admin.NotMatchAdminPasswordException;
 import ddangkong.facade.BaseServiceTest;
+import ddangkong.facade.admin.dto.AdminLoginRequest;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,18 +24,21 @@ class AdminServiceTest extends BaseServiceTest {
 
         @Test
         void 어드민_비밀번호가_일치하면_아무_일도_일어나지_않는다() {
+            // given
+            AdminLoginRequest request = new AdminLoginRequest("admin", adminPassword);
+
             // when & then
-            assertThatCode(() -> adminService.validatePassword(adminPassword))
+            assertThatCode(() -> adminService.validatePassword(request))
                     .doesNotThrowAnyException();
         }
 
         @Test
         void 어드민_비밀번호가_다르면_예외가_발생한다() {
             // given
-            String notMatchPassword = "no-password";
+            AdminLoginRequest request = new AdminLoginRequest("admin", "no-password");
 
             // when & then
-            assertThatThrownBy(() -> adminService.validatePassword(notMatchPassword))
+            assertThatThrownBy(() -> adminService.validatePassword(request))
                     .isInstanceOf(NotMatchAdminPasswordException.class);
         }
     }
